+++ demos/eslint.config.mjs
@@ -1,0 +1,11 @@
+// @ts-nocheck
+
+import tseslint from "typescript-eslint";
+import baseConfig from "../eslint.config.mjs";
+
+export default tseslint.config(baseConfig, {
+  rules: {
+    "@typescript-eslint/no-unsafe-argument": "off",
+    "@typescript-eslint/no-unsafe-assignment": "off",
+  },
+});